svg/text/parsing/inline-size-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing inline-size with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#InlineSizeProperty" rel="help"/><h:link href="https://drafts.csswg.org/css-logical/#dimension-properties" rel="help"/><h:meta content="inline-size supports only the grammar 'auto | length-percentage'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("inline-size", "none");
test_invalid_value("inline-size", "-10px");
test_invalid_value("inline-size", "20");
test_invalid_value("inline-size", "30deg");
test_invalid_value("inline-size", "40px 50%");

  ]]></script></svg>
